Move associate um network printer function into UM3OutputDevicePlugin
