+++ plugins/UM3NetworkPrinting/src/DiscoverUM3Action.py
@@ -124,43 +124,8 @@ class DiscoverUM3Action(MachineAction):
     # stored into the metadata of the currently active machine.
     @pyqtSlot(QObject)
     def associateActiveMachineWithPrinterDevice(self, printer_device: Optional["PrinterOutputDevice"]) -> None:
-        if not printer_device:
-            return
-
-        Logger.log("d", "Attempting to set the network key of the active machine to %s", printer_device.key)
-
-        global_container_stack = CuraApplication.getInstance().getGlobalContainerStack()
-        if not global_container_stack:
-            return
-
-        meta_data = global_container_stack.getMetaData()
-
-        if "um_network_key" in meta_data:  # Global stack already had a connection, but it's changed.
-            old_network_key = meta_data["um_network_key"]
-            # Since we might have a bunch of hidden stacks, we also need to change it there.
-            metadata_filter = {"um_network_key": old_network_key}
-            containers = CuraContainerRegistry.getInstance().findContainerStacks(type="machine", **metadata_filter)
-
-            for container in containers:
-                container.setMetaDataEntry("um_network_key", printer_device.key)
-
-                # Delete old authentication data.
-                Logger.log("d", "Removing old authentication id %s for device %s",
-                           global_container_stack.getMetaDataEntry("network_authentication_id", None), printer_device.key)
-
-                container.removeMetaDataEntry("network_authentication_id")
-                container.removeMetaDataEntry("network_authentication_key")
-
-                # Ensure that these containers do know that they are configured for network connection
-                container.addConfiguredConnectionType(printer_device.connectionType.value)
-
-        else:  # Global stack didn't have a connection yet, configure it.
-            global_container_stack.setMetaDataEntry("um_network_key", printer_device.key)
-            global_container_stack.addConfiguredConnectionType(printer_device.connectionType.value)
-
         if self._network_plugin:
-            # Ensure that the connection states are refreshed.
-            self._network_plugin.refreshConnections()
+            self._network_plugin.associateActiveMachineWithPrinterDevice(printer_device)
 
     @pyqtSlot(result = str)
     def getStoredKey(self) -> str:
